requirements.txt
python-dotenv
streamlit
requests
beautifulsoup4
openai
langchain
langchain-openai
chromadb